[Feat]SpotPage의 components들을 page에서 component로 옮김
