+++ src/layouts/MapLayout.jsx
@@ -6,7 +6,6 @@ import {
   useNavermaps,
 } from 'react-naver-maps';
 import Header from '../components/Header';
-import Footer from '../components/Footer';
 
 export default function MapLayout() {
   const navermaps = useNavermaps();
@@ -16,9 +15,10 @@ export default function MapLayout() {
       <Outlet /> {/* 자식 컴포넌트가 렌더링될 위치 */}
       {/* 변경되지 않을 컴포넌트 */}
       <MapDiv
+        className="overflow-hidden"
         style={{
           width: '100%',
-          height: '600px',
+          height: '680px',
         }}>
         <NaverMap
           defaultCenter={new navermaps.LatLng(37.3595704, 127.105399)}
@@ -27,7 +27,6 @@ export default function MapLayout() {
             defaultPosition={new navermaps.LatLng(37.3595704, 127.105399)} />
         </NaverMap>{' '}
       </MapDiv>
-      <Footer />
     </div>
   );
 };
